refactor: package 변경 - view dir
