+++ nextstep-mvc/src/main/java/nextstep/mvc/asis/Controller.java
@@ -1,8 +1,8 @@
 package nextstep.mvc.asis;
 
 import nextstep.mvc.tobe.HandlerResolver;
-import nextstep.mvc.tobe.JspView;
-import nextstep.mvc.tobe.ModelAndView;
+import nextstep.mvc.tobe.view.JspView;
+import nextstep.mvc.tobe.view.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
