+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/AnnotationHandlerMapping.java
@@ -2,6 +2,8 @@ package nextstep.mvc.tobe;
 
 import com.google.common.collect.Maps;
 import nextstep.mvc.HandlerMapping;
+import nextstep.mvc.tobe.view.JspViewResolver;
+import nextstep.mvc.tobe.view.ViewResolver;
 import nextstep.web.annotation.Controller;
 import nextstep.web.annotation.RequestMapping;
 import nextstep.web.annotation.RequestMethod;
