+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/view/JspViewResolver.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.tobe.view;
 
 public class JspViewResolver implements ViewResolver {
     @Override
