+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/view/ViewResolver.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.tobe.view;
 
 public interface ViewResolver {
     View resolve(Object result);
